Dark-Blue Bot/requirements.txt
discord
pyshorteners
praw
flask
google
google-cloud-speech
gtts
"discord.py"